fix projects modal and optimize

+++ client/src/components/clients/ClientRow.tsx
@@ -24,15 +24,18 @@ const ClientRow: React.FC<ClientRowProps> = (props) => {
             <TableCell>{props.client?.name}</TableCell>
             <TableCell>{props.client?.email}</TableCell>
             <TableCell>{props.client?.phone}</TableCell>
-            <TableCell align={"right"}>
+            <TableCell align={"right"} sx={{display: 'flex'}}>
                 <Box sx={{display: 'flex', gap: 1}}>
                     <Button variant={'contained'} size={'small'} color={'primary'} onClick={() => {
                         props.handleOpen(props.client?.id as string ?? "")
                     }}>
                         <ModeEditOutlinedIcon/>
                     </Button>
-                    <Button variant={'contained'} size={'small'} color={'error'} onClick={() => {
-                        deleteClients()
+                    <Button variant={'contained'} size={'small'} color={'error'} onClick={async () => {
+                        const result = confirm('Do you want to delete this client id' + props?.client?.id)
+                        if (result) {
+                            await deleteClients()
+                        }
                     }}>
                         <DeleteOutlineOutlinedIcon/>
                     </Button>
